Added AbstractCollapser
