+++ hystrix-core/src/main/java/com/netflix/hystrix/HystrixObservableCollapser.java
@@ -16,36 +16,20 @@
 package com.netflix.hystrix;
 
 import com.netflix.hystrix.HystrixCollapser.CollapsedRequest;
-import com.netflix.hystrix.HystrixCommandProperties.ExecutionIsolationStrategy;
 import com.netflix.hystrix.collapser.CollapserTimer;
-import com.netflix.hystrix.collapser.HystrixCollapserBridge;
 import com.netflix.hystrix.collapser.RealCollapserTimer;
-import com.netflix.hystrix.collapser.RequestCollapser;
 import com.netflix.hystrix.collapser.RequestCollapserFactory;
-import com.netflix.hystrix.strategy.HystrixPlugins;
 import com.netflix.hystrix.strategy.concurrency.HystrixRequestContext;
-import com.netflix.hystrix.strategy.metrics.HystrixMetricsPublisherFactory;
-import com.netflix.hystrix.strategy.properties.HystrixPropertiesFactory;
 import com.netflix.hystrix.strategy.properties.HystrixPropertiesStrategy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import rx.Observable;
-import rx.Scheduler;
-import rx.Subscription;
 import rx.functions.Action0;
 import rx.functions.Action1;
 import rx.functions.Func0;
 import rx.functions.Func1;
-import rx.schedulers.Schedulers;
-import rx.subjects.ReplaySubject;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.*;
 
 /**
  * Collapse multiple requests into a single {@link HystrixCommand} execution based on a time window and optionally a max batch size.
@@ -68,14 +52,9 @@ import java.util.concurrent.ConcurrentHashMap;
  * @param <RequestArgumentType>
  *            The type of the request argument. If multiple arguments are needed, wrap them in another object or a Tuple.
  */
-public abstract class HystrixObservableCollapser<K, BatchReturnType, ResponseType, RequestArgumentType> implements HystrixObservable<ResponseType> {
+public abstract class HystrixObservableCollapser<K, BatchReturnType, ResponseType, RequestArgumentType> extends AbstractCollapser<BatchReturnType, ResponseType, RequestArgumentType> {
 
     static final Logger logger = LoggerFactory.getLogger(HystrixObservableCollapser.class);
-
-    private final RequestCollapserFactory<BatchReturnType, ResponseType, RequestArgumentType> collapserFactory;
-    private final HystrixRequestCache requestCache;
-    private final HystrixCollapserBridge<BatchReturnType, ResponseType, RequestArgumentType> collapserInstanceWrapper;
-    private final HystrixCollapserMetrics metrics;
 
     /**
      * The scope of request collapsing.
@@ -122,119 +101,73 @@ public abstract class HystrixObservableCollapser<K, BatchReturnType, ResponseTyp
     }
 
     /* package for tests */HystrixObservableCollapser(HystrixCollapserKey collapserKey, Scope scope, CollapserTimer timer, HystrixCollapserProperties.Setter propertiesBuilder, HystrixCollapserMetrics metrics) {
-        if (collapserKey == null || collapserKey.name().trim().equals("")) {
-            String defaultKeyName = getDefaultNameFromClass(getClass());
-            collapserKey = HystrixCollapserKey.Factory.asKey(defaultKeyName);
+        super(collapserKey, scope, timer, propertiesBuilder, metrics);
+    }
+
+    @Override
+    public Scope getScope() {
+        return Scope.valueOf(super.getScope().name());
+    }
+
+    @Override
+    protected Observable<Void> handleResponse(Observable<BatchReturnType> batchResponse, Collection<CollapsedRequest<ResponseType, RequestArgumentType>> requests) {
+        Func1<RequestArgumentType, K> requestKeySelector = getRequestArgumentKeySelector();
+        final Func1<BatchReturnType, K> batchResponseKeySelector = getBatchReturnTypeKeySelector();
+        final Func1<BatchReturnType, ResponseType> mapBatchTypeToResponseType = getBatchReturnTypeToResponseTypeMapper();
+
+        // index the requests by key
+        final Map<K, CollapsedRequest<ResponseType, RequestArgumentType>> requestsByKey = new HashMap<K, CollapsedRequest<ResponseType, RequestArgumentType>>(requests.size());
+        for (CollapsedRequest<ResponseType, RequestArgumentType> cr : requests) {
+            K requestArg = requestKeySelector.call(cr.getArgument());
+            requestsByKey.put(requestArg, cr);
         }
+        final Set<K> seenKeys = new HashSet<K>();
 
-        HystrixCollapserProperties properties = HystrixPropertiesFactory.getCollapserProperties(collapserKey, propertiesBuilder);
-        this.collapserFactory = new RequestCollapserFactory<BatchReturnType, ResponseType, RequestArgumentType>(collapserKey, scope, timer, properties);
-        this.requestCache = HystrixRequestCache.getInstance(collapserKey, HystrixPlugins.getInstance().getConcurrencyStrategy());
-
-        if (metrics == null) {
-            this.metrics = HystrixCollapserMetrics.getInstance(collapserKey, properties);
-        } else {
-            this.metrics = metrics;
-        }
-
-        final HystrixObservableCollapser<K, BatchReturnType, ResponseType, RequestArgumentType> self = this;
-
-           /* strategy: HystrixMetricsPublisherCollapser */
-        HystrixMetricsPublisherFactory.createOrRetrievePublisherForCollapser(collapserKey, this.metrics, properties);
-
-
-        /**
-         * Used to pass public method invocation to the underlying implementation in a separate package while leaving the methods 'protected' in this class.
-         */
-        collapserInstanceWrapper = new HystrixCollapserBridge<BatchReturnType, ResponseType, RequestArgumentType>() {
-
+        // observe the responses and join with the requests by key
+        return batchResponse.doOnNext(new Action1<BatchReturnType>() {
             @Override
-            public Collection<Collection<CollapsedRequest<ResponseType, RequestArgumentType>>> shardRequests(Collection<CollapsedRequest<ResponseType, RequestArgumentType>> requests) {
-                Collection<Collection<CollapsedRequest<ResponseType, RequestArgumentType>>> shards = self.shardRequests(requests);
-                self.metrics.markShards(shards.size());
-                return shards;
-            }
-
-            @Override
-            public Observable<BatchReturnType> createObservableCommand(Collection<CollapsedRequest<ResponseType, RequestArgumentType>> requests) {
-                HystrixObservableCommand<BatchReturnType> command = self.createCommand(requests);
-
-                // mark the number of requests being collapsed together
-                command.markAsCollapsedCommand(this.getCollapserKey(), requests.size());
-                self.metrics.markBatch(requests.size());
-                return command.toObservable();
-            }
-
-            @Override
-            public Observable<Void> mapResponseToRequests(Observable<BatchReturnType> batchResponse, Collection<CollapsedRequest<ResponseType, RequestArgumentType>> requests) {
-                Func1<RequestArgumentType, K> requestKeySelector = self.getRequestArgumentKeySelector();
-                final Func1<BatchReturnType, K> batchResponseKeySelector = self.getBatchReturnTypeKeySelector();
-                final Func1<BatchReturnType, ResponseType> mapBatchTypeToResponseType = self.getBatchReturnTypeToResponseTypeMapper();
-
-                // index the requests by key
-                final Map<K, CollapsedRequest<ResponseType, RequestArgumentType>> requestsByKey = new HashMap<K, CollapsedRequest<ResponseType, RequestArgumentType>>(requests.size());
-                for (CollapsedRequest<ResponseType, RequestArgumentType> cr : requests) {
-                    K requestArg = requestKeySelector.call(cr.getArgument());
-                    requestsByKey.put(requestArg, cr);
+            public void call(BatchReturnType batchReturnType) {
+                try {
+                    K responseKey = batchResponseKeySelector.call(batchReturnType);
+                    CollapsedRequest<ResponseType, RequestArgumentType> requestForResponse = requestsByKey.get(responseKey);
+                    if (requestForResponse != null) {
+                        requestForResponse.emitResponse(mapBatchTypeToResponseType.call(batchReturnType));
+                        // now add this to seenKeys, so we can later check what was seen, and what was unseen
+                        seenKeys.add(responseKey);
+                    } else {
+                        logger.warn("Batch Response contained a response key not in request batch : {}", responseKey);
+                    }
+                } catch (Throwable ex) {
+                    logger.warn("Uncaught error during demultiplexing of BatchResponse", ex);
                 }
-                final Set<K> seenKeys = new HashSet<K>();
-
-                // observe the responses and join with the requests by key
-                return batchResponse
-                        .doOnNext(new Action1<BatchReturnType>() {
-                            @Override
-                            public void call(BatchReturnType batchReturnType) {
-                                try {
-                                    K responseKey = batchResponseKeySelector.call(batchReturnType);
-                                    CollapsedRequest<ResponseType, RequestArgumentType> requestForResponse = requestsByKey.get(responseKey);
-                                    if (requestForResponse != null) {
-                                        requestForResponse.emitResponse(mapBatchTypeToResponseType.call(batchReturnType));
-                                        // now add this to seenKeys, so we can later check what was seen, and what was unseen
-                                        seenKeys.add(responseKey);
-                                    } else {
-                                        logger.warn("Batch Response contained a response key not in request batch : {}", responseKey);
-                                    }
-                                } catch (Throwable ex) {
-                                    logger.warn("Uncaught error during demultiplexing of BatchResponse", ex);
-                                }
-                            }
-                        })
-                        .doOnError(new Action1<Throwable>() {
-                            @Override
-                            public void call(Throwable t) {
-                                Exception ex = getExceptionFromThrowable(t);
-                                for (CollapsedRequest<ResponseType, RequestArgumentType> collapsedReq : requestsByKey.values()) {
-                                    collapsedReq.setException(ex);
-                                }
-                            }
-                        })
-                        .doOnCompleted(new Action0() {
-                            @Override
-                            public void call() {
-
-                                for (Map.Entry<K, CollapsedRequest<ResponseType, RequestArgumentType>> entry : requestsByKey.entrySet()) {
-                                    K key = entry.getKey();
-                                    CollapsedRequest<ResponseType, RequestArgumentType> collapsedReq = entry.getValue();
-                                    if (!seenKeys.contains(key)) {
-                                        try {
-                                            onMissingResponse(collapsedReq);
-                                        } catch (Throwable ex) {
-                                            collapsedReq.setException(new RuntimeException("Error in HystrixObservableCollapser.onMissingResponse handler", ex));
-                                        }
-                                    }
-                                    //then unconditionally issue an onCompleted. this ensures the downstream gets a terminal, regardless of how onMissingResponse was implemented
-                                    collapsedReq.setComplete();
-                                }
-                            }
-                        }).ignoreElements().cast(Void.class);
             }
-
+        }).doOnError(new Action1<Throwable>() {
             @Override
-            public HystrixCollapserKey getCollapserKey() {
-                return self.getCollapserKey();
+            public void call(Throwable t) {
+                Exception ex = getExceptionFromThrowable(t);
+                for (CollapsedRequest<ResponseType, RequestArgumentType> collapsedReq : requestsByKey.values()) {
+                    collapsedReq.setException(ex);
+                }
             }
+        }).doOnCompleted(new Action0() {
+            @Override
+            public void call() {
 
-        };
+                for (Map.Entry<K, CollapsedRequest<ResponseType, RequestArgumentType>> entry : requestsByKey.entrySet()) {
+                    K key = entry.getKey();
+                    CollapsedRequest<ResponseType, RequestArgumentType> collapsedReq = entry.getValue();
+                    if (!seenKeys.contains(key)) {
+                        try {
+                            onMissingResponse(collapsedReq);
+                        } catch (Throwable ex) {
+                            collapsedReq.setException(new RuntimeException("Error in HystrixObservableCollapser.onMissingResponse handler", ex));
+                        }
+                    }
+                    //then unconditionally issue an onCompleted. this ensures the downstream gets a terminal, regardless of how onMissingResponse was implemented
+                    collapsedReq.setComplete();
+                }
+            }
+        }).ignoreElements().cast(Void.class);
     }
 
     protected Exception getExceptionFromThrowable(Throwable t) {
@@ -246,95 +179,6 @@ public abstract class HystrixObservableCollapser<K, BatchReturnType, ResponseTyp
             e = new Exception("Throwable caught while executing.", t);
         }
         return e;
-    }
-
-
-    private HystrixCollapserProperties getProperties() {
-        return collapserFactory.getProperties();
-    }
-
-    /**
-     * Key of the {@link HystrixObservableCollapser} used for properties, metrics, caches, reporting etc.
-     * 
-     * @return {@link HystrixCollapserKey} identifying this {@link HystrixObservableCollapser} instance
-     */
-    public HystrixCollapserKey getCollapserKey() {
-        return collapserFactory.getCollapserKey();
-    }
-
-    /**
-     * Scope of collapsing.
-     * <p>
-     * <ul>
-     * <li>REQUEST: Requests within the scope of a {@link HystrixRequestContext} will be collapsed.
-     * <p>
-     * Typically this means that requests within a single user-request (ie. HTTP request) are collapsed. No interaction with other user requests. 1 queue per user request.
-     * </li>
-     * <li>GLOBAL: Requests from any thread (ie. all HTTP requests) within the JVM will be collapsed. 1 queue for entire app.</li>
-     * </ul>
-     * <p>
-     * Default: {@link Scope#REQUEST} (defined via constructor)
-     * 
-     * @return {@link Scope} that collapsing should be performed within.
-     */
-    public Scope getScope() {
-        return Scope.valueOf(collapserFactory.getScope().name());
-    }
-
-    /**
-     * Return the {@link HystrixCollapserMetrics} for this collapser
-     * @return {@link HystrixCollapserMetrics} for this collapser
-     */
-    public HystrixCollapserMetrics getMetrics() {
-        return metrics;
-    }
-
-    /**
-     * The request arguments to be passed to the {@link HystrixCommand}.
-     * <p>
-     * Typically this means to take the argument(s) provided to the constructor and return it here.
-     * <p>
-     * If there are multiple arguments that need to be bundled, create a single object to contain them, or use a Tuple.
-     * 
-     * @return RequestArgumentType
-     */
-    public abstract RequestArgumentType getRequestArgument();
-
-    /**
-     * Factory method to create a new {@link HystrixObservableCommand}{@code <BatchReturnType>} command object each time a batch needs to be executed.
-     * <p>
-     * Do not return the same instance each time. Return a new instance on each invocation.
-     * <p>
-     * Process the 'requests' argument into the arguments the command object needs to perform its work.
-     * <p>
-     * If a batch or requests needs to be split (sharded) into multiple commands, see {@link #shardRequests} <p>
-     * IMPLEMENTATION NOTE: Be fast (ie. <1ms) in this method otherwise it can block the Timer from executing subsequent batches. Do not do any processing beyond constructing the command and returning
-     * it.
-     * 
-     * @param requests
-     *            {@code Collection<CollapsedRequest<ResponseType, RequestArgumentType>>} containing {@link CollapsedRequest} objects containing the arguments of each request collapsed in this batch.
-     * @return {@link HystrixObservableCommand}{@code <BatchReturnType>} which when executed will retrieve results for the batch of arguments as found in the Collection of {@link CollapsedRequest}
-     *         objects
-     */
-    protected abstract HystrixObservableCommand<BatchReturnType> createCommand(Collection<CollapsedRequest<ResponseType, RequestArgumentType>> requests);
-
-    /**
-     * Override to split (shard) a batch of requests into multiple batches that will each call <code>createCommand</code> separately.
-     * <p>
-     * The purpose of this is to allow collapsing to work for services that have sharded backends and batch executions that need to be shard-aware.
-     * <p>
-     * For example, a batch of 100 requests could be split into 4 different batches sharded on name (ie. a-g, h-n, o-t, u-z) that each result in a separate {@link HystrixCommand} being created and
-     * executed for them.
-     * <p>
-     * By default this method does nothing to the Collection and is a pass-thru.
-     * 
-     * @param requests
-     *            {@code Collection<CollapsedRequest<ResponseType, RequestArgumentType>>} containing {@link CollapsedRequest} objects containing the arguments of each request collapsed in this batch.
-     * @return Collection of {@code Collection<CollapsedRequest<ResponseType, RequestArgumentType>>} objects sharded according to business rules.
-     *         <p>The CollapsedRequest instances should not be modified or wrapped as the CollapsedRequest instance object contains state information needed to complete the execution.
-     */
-    protected Collection<Collection<CollapsedRequest<ResponseType, RequestArgumentType>>> shardRequests(Collection<CollapsedRequest<ResponseType, RequestArgumentType>> requests) {
-        return Collections.singletonList(requests);
     }
 
     /**
@@ -373,151 +217,6 @@ public abstract class HystrixObservableCollapser<K, BatchReturnType, ResponseTyp
      * @return function for mapping from BatchReturnType to ResponseType
      */
     protected abstract Func1<BatchReturnType, ResponseType> getBatchReturnTypeToResponseTypeMapper();
-
-    /**
-     * Used for asynchronous execution with a callback by subscribing to the {@link Observable}.
-     * <p>
-     * This eagerly starts execution the same as {@link HystrixCollapser#queue()} and {@link HystrixCollapser#execute()}.
-     * A lazy {@link Observable} can be obtained from {@link #toObservable()}.
-     * <p>
-     * <b>Callback Scheduling</b>
-     * <p>
-     * <ul>
-     * <li>When using {@link ExecutionIsolationStrategy#THREAD} this defaults to using {@link Schedulers#computation()} for callbacks.</li>
-     * <li>When using {@link ExecutionIsolationStrategy#SEMAPHORE} this defaults to using {@link Schedulers#immediate()} for callbacks.</li>
-     * </ul>
-     * Use {@link #toObservable(rx.Scheduler)} to schedule the callback differently.
-     * <p>
-     * See https://github.com/Netflix/RxJava/wiki for more information.
-     * 
-     * @return {@code Observable<R>} that executes and calls back with the result of of {@link HystrixCommand}{@code <BatchReturnType>} execution after mapping
-     *         the {@code <BatchReturnType>} into {@code <ResponseType>}
-     */
-    public Observable<ResponseType> observe() {
-        // use a ReplaySubject to buffer the eagerly subscribed-to Observable
-        ReplaySubject<ResponseType> subject = ReplaySubject.create();
-        // eagerly kick off subscription
-        final Subscription underlyingSubscription = toObservable().subscribe(subject);
-        // return the subject that can be subscribed to later while the execution has already started
-        return subject.doOnUnsubscribe(new Action0() {
-            @Override
-            public void call() {
-                underlyingSubscription.unsubscribe();
-            }
-        });
-    }
-
-    /**
-     * A lazy {@link Observable} that will execute when subscribed to.
-     * <p>
-     * <b>Callback Scheduling</b>
-     * <p>
-     * <ul>
-     * <li>When using {@link ExecutionIsolationStrategy#THREAD} this defaults to using {@link Schedulers#computation()} for callbacks.</li>
-     * <li>When using {@link ExecutionIsolationStrategy#SEMAPHORE} this defaults to using {@link Schedulers#immediate()} for callbacks.</li>
-     * </ul>
-     * <p>
-     * See https://github.com/Netflix/RxJava/wiki for more information.
-     * 
-     * @return {@code Observable<R>} that lazily executes and calls back with the result of of {@link HystrixCommand}{@code <BatchReturnType>} execution after mapping the
-     * {@code <BatchReturnType>} into {@code <ResponseType>}
-     */
-    public Observable<ResponseType> toObservable() {
-        // when we callback with the data we want to do the work
-        // on a separate thread than the one giving us the callback
-        return toObservable(Schedulers.computation());
-    }
-
-    /**
-     * A lazy {@link Observable} that will execute when subscribed to.
-     * <p>
-     * See https://github.com/Netflix/RxJava/wiki for more information.
-     * 
-     * @param observeOn
-     *            The {@link Scheduler} to execute callbacks on.
-     * @return {@code Observable<R>} that lazily executes and calls back with the result of of {@link HystrixCommand}{@code <BatchReturnType>} execution after mapping the
-     * {@code <BatchReturnType>} into {@code <ResponseType>}
-     */
-    public Observable<ResponseType> toObservable(Scheduler observeOn) {
-
-        return Observable.defer(new Func0<Observable<ResponseType>>() {
-            @Override
-            public Observable<ResponseType> call() {
-                final boolean isRequestCacheEnabled = getProperties().requestCacheEnabled().get();
-
-                /* try from cache first */
-                if (isRequestCacheEnabled) {
-                    HystrixCachedObservable<ResponseType> fromCache = requestCache.get(getCacheKey());
-                    if (fromCache != null) {
-                        metrics.markResponseFromCache();
-                        return fromCache.toObservable();
-                    }
-                }
-
-                RequestCollapser<BatchReturnType, ResponseType, RequestArgumentType> requestCollapser = collapserFactory.getRequestCollapser(collapserInstanceWrapper);
-                Observable<ResponseType> response = requestCollapser.submitRequest(getRequestArgument());
-                metrics.markRequestBatched();
-                if (isRequestCacheEnabled) {
-                    /*
-                     * A race can occur here with multiple threads queuing but only one will be cached.
-                     * This means we can have some duplication of requests in a thread-race but we're okay
-                     * with having some inefficiency in duplicate requests in the same batch
-                     * and then subsequent requests will retrieve a previously cached Observable.
-                     *
-                     * If this is an issue we can make a lazy-future that gets set in the cache
-                     * then only the winning 'put' will be invoked to actually call 'submitRequest'
-                     */
-                    HystrixCachedObservable<ResponseType> toCache = HystrixCachedObservable.from(response);
-                    HystrixCachedObservable<ResponseType> fromCache = requestCache.putIfAbsent(getCacheKey(), toCache);
-                    if (fromCache == null) {
-                        return toCache.toObservable();
-                    } else {
-                        return fromCache.toObservable();
-                    }
-                }
-                return response;
-            }
-        });
-    }
-
-    /**
-     * Key to be used for request caching.
-     * <p>
-     * By default this returns null which means "do not cache".
-     * <p>
-     * To enable caching override this method and return a string key uniquely representing the state of a command instance.
-     * <p>
-     * If multiple command instances in the same request scope match keys then only the first will be executed and all others returned from cache.
-     * 
-     * @return String cacheKey or null if not to cache
-     */
-    protected String getCacheKey() {
-        return null;
-    }
-
-    /**
-     * Clears all state. If new requests come in instances will be recreated and metrics started from scratch.
-     */
-    /* package */static void reset() {
-        RequestCollapserFactory.reset();
-    }
-
-    private static String getDefaultNameFromClass(@SuppressWarnings("rawtypes") Class<? extends HystrixObservableCollapser> cls) {
-        String fromCache = defaultNameCache.get(cls);
-        if (fromCache != null) {
-            return fromCache;
-        }
-        // generate the default
-        // default HystrixCommandKey to use if the method is not overridden
-        String name = cls.getSimpleName();
-        if (name.equals("")) {
-            // we don't have a SimpleName (anonymous inner class) so use the full class name
-            name = cls.getName();
-            name = name.substring(name.lastIndexOf('.') + 1, name.length());
-        }
-        defaultNameCache.put(cls, name);
-        return name;
-    }
 
     /**
      * Fluent interface for arguments to the {@link HystrixObservableCollapser} constructor.
@@ -581,10 +280,4 @@ public abstract class HystrixObservableCollapser<K, BatchReturnType, ResponseTyp
         }
 
     }
-
-    // this is a micro-optimization but saves about 1-2microseconds (on 2011 MacBook Pro) 
-    // on the repetitive string processing that will occur on the same classes over and over again
-    @SuppressWarnings("rawtypes")
-    private static ConcurrentHashMap<Class<? extends HystrixObservableCollapser>, String> defaultNameCache = new ConcurrentHashMap<Class<? extends HystrixObservableCollapser>, String>();
-
 }
